requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
ecdsa==0.18.0
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
